refactor: rename openaicompat to relayconvert for improved clarity
