+++ .gitignore
@@ -39,4 +39,4 @@ skills-lock.json
 
 # Local-only live probes and scratch test workspaces.
 .local-tests/
-service/openaicompat/chat_responses_live_local_test.go
+service/relayconvert/chat_responses_live_local_test.go
