+++ service/openai_chat_responses_mode.go
@@ -1,14 +1,14 @@
 package service
 
 import (
-	"github.com/QuantumNous/new-api/service/openaicompat"
+	"github.com/QuantumNous/new-api/service/relayconvert"
 	"github.com/QuantumNous/new-api/setting/model_setting"
 )
 
 func ShouldChatCompletionsUseResponsesPolicy(policy model_setting.ChatCompletionsToResponsesPolicy, channelID int, channelType int, model string) bool {
-	return openaicompat.ShouldChatCompletionsUseResponsesPolicy(policy, channelID, channelType, model)
+	return relayconvert.ShouldChatCompletionsUseResponsesPolicy(policy, channelID, channelType, model)
 }
 
 func ShouldChatCompletionsUseResponsesGlobal(channelID int, channelType int, model string) bool {
-	return openaicompat.ShouldChatCompletionsUseResponsesGlobal(channelID, channelType, model)
+	return relayconvert.ShouldChatCompletionsUseResponsesGlobal(channelID, channelType, model)
 }
