+++ service/relayconvert/chat_responses_compat_test.go
@@ -1,4 +1,4 @@
-package openaicompat
+package relayconvert
 
 import (
 	"testing"
