+++ service/relayconvert/chat_to_responses.go
@@ -1,4 +1,4 @@
-package openaicompat
+package relayconvert
 
 import (
 	"encoding/json"
