+++ service/relayconvert/responses_to_chat.go
@@ -1,4 +1,4 @@
-package openaicompat
+package relayconvert
 
 import (
 	"errors"
